habit_tracker/pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "requests (>=2.33.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "habit_tracker", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"